+++ resources/views/back/managers/create.blade.php
@@ -43,6 +43,25 @@
 
 @stop
 
+
+
+@push('custom-scripts-after-vue')
+
+    <script>
+        // Display a warning message to tell the user that adding images to redactor is only possible after page creation.
+        var editors = document.querySelectorAll('[data-editor]');
+        for(var i = 0; i < editors.length; i++) {
+            var message = document.createElement('p');
+            message.style = "font-style: italic; opacity: 0.3; margin-top: 1rem;";
+            message.innerHTML = "Het toevoegen van afbeeldingen aan de wysiwyg is pas mogelijk na het aanmaken van dit item.";
+            editors[i].parentElement.appendChild(message);
+        }
+    </script>
+
+    @include('chief::back._layouts._partials.editor-script', ['disableImageUpload' => true])
+
+@endpush
+
 @include('chief::back._elements.file-component')
 @include('chief::back._elements.slimcropper-component')
 @include('chief::back._elements.fileupload-component')