+++ src/HealthMonitor/Checks/HomepageAccessibleCheck.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 
 namespace Thinktomorrow\Chief\HealthMonitor\Checks;
 
+use Thinktomorrow\Chief\HealthMonitor\Notifiers\AlertBarNotifier;
 use Thinktomorrow\Chief\Settings\Homepage;
 
 class HomepageAccessibleCheck implements HealthCheck
@@ -34,5 +35,12 @@ class HomepageAccessibleCheck implements HealthCheck
     public function message(): string
     {
         return 'Het lijkt erop dat de homepagina niet meer bereikbaar is. <a href="'. route('chief.back.settings.edit') .'" class="text-secondary-800 underline hover:text-white">Kies een nieuwe</a>.';
+    }
+
+    public function notifiers(): array
+    {
+        return [
+            AlertBarNotifier::class,
+        ];
     }
 }